requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
configparser==4.0.2
cryptography==3.4.7
discord.py==1.7.3
idna==3.2
multidict==5.1.0
oci==2.41.1
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
tabulate==0.8.9
typing-extensions==3.10.0.0
yarl==1.6.3
